requirements.txt
opencv_python_headless==4.8.1.78
numpy==2.2.4
Pillow==11.1.0
